integrted script with vscode extension
